# Speedup encoder